+++ app/src/main/java/com/example/rgb/database/categories/MacroCategoryEntity.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 data class MacroCategoryEntity(
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "macroCategoryId")
-    val macroCategoryId: Int = 0,
+    val macroCategoryId: Long = 0,
 
     @ColumnInfo(name = "macroCategoryName")
     val macroCategoryName: String
